third_party/blink/web_tests/editing/input/setting-input-value-cancel-ime-composition-expected.txt

This test ensures WebKit clears IME composition inside an input element when the value is programmatically set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

nihao should marked from 0 to 5
PASS textInputController.hasMarkedText() is true
PASS textInputController.markedRange().toString() is "0,5"
text was removed so there should be no marked text
PASS textInputController.hasMarkedText() is false
nihao should marked from 0 to 5
PASS textInputController.hasMarkedText() is true
PASS textInputController.markedRange().toString() is "0,5"
the text was changed so there should be no marked text
PASS textInputController.hasMarkedText() is false
PASS successfullyParsed is true

TEST COMPLETE
